requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
psutil==5.9.6
aiofiles==23.2.1
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0